analyzer/app/requirements.txt
grpcio
grpcio-tools
librosa
soundfile
numpy<2